Keep exceptionCatchers separate from callbacks
This fixes the issue noted by isaacs here: http://gist.github.com/283562
in which because the exception catcher is kept as a property of the
callback function, it can be overwritten if a function is registered
as a callback for multiple events.

Previously we had:

eventEmitter._events["eventname"][n] is the callback
eventEmitter._events["eventname"][n]._exceptionCatcher is um, the
exception catcher, attached as a property of the callback function.

This now becomes:

eventEmitter._events["eventname"].callbacks[n] is the callback
eventEmitter._events["eventname"].catchers[n] is the exception catcher

This way we retain the current behaviour of
EventEmitter.listeners() returning an array of callbacks, and
EventEmitter.removeListener() using a simple indexOf search.